requirements.txt
selenium
beautifulsoup4
requests
schedule
python-dotenv
SQLAlchemy
